💥 Working ROS 2 Control - tested on RG2

+++ include/onrobot_driver/RG.hpp
@@ -9,7 +9,7 @@
 #include "onrobot_driver/TCPConnectionWrapper.hpp"
 #include "onrobot_driver/SerialConnectionWrapper.hpp"
 
-// Modbus request/response definitions and utilities.
+// Modbus request/response definitions and utilities
 #include "MB/modbusRequest.hpp"
 #include "MB/modbusResponse.hpp"
 #include "MB/modbusException.hpp"
@@ -17,7 +17,7 @@
 
 class RG {
 public:
-    // Constructors for TCP and Serial connections.
+    // Constructors for TCP and Serial connections
     RG(const std::string &type, const std::string &ip, int port);
     RG(const std::string &type, const std::string &device);
     ~RG();
@@ -29,26 +29,27 @@ public:
     float getWidthWithOffset();
 
     // Write commands
-    void setFingertipOffset(uint16_t offset_val);
-    void setTargetForce(uint16_t force_val);
-    void setTargetWidth(uint16_t width_val);
+    void setFingertipOffset(float offset_val);
+    void setTargetForce(float force_val);
+    void setTargetWidth(float width_val);
     void setControlMode(uint16_t command);
     
+    // Gripper control commands
     void closeGripper();
     void openGripper();
-    void moveGripper(uint16_t width_val);
+    void moveGripper(float width_val);
 
 private:
     std::unique_ptr<IModbusConnection> connection;
     std::string type;
-    uint16_t max_width;
-    uint16_t max_force;
+    float max_width;
+    float max_force;
 
-    // Default parameters.
-    uint16_t default_force;
-    uint16_t default_fingertip_offset;
+    // Default parameters
+    float default_force;
+    float default_fingertip_offset;
 
-    // Constants for registers and commands.
+    // Constants for registers and commands
     static constexpr uint16_t DEVICE_ID = 65;
     static constexpr uint16_t REG_TARGET_FORCE = 0;
     static constexpr uint16_t REG_TARGET_WIDTH = 1;
@@ -61,11 +62,11 @@ private:
     static constexpr uint16_t REG_ACTUAL_WIDTH_WITH_OFFSET = 275;
     static constexpr uint16_t REG_SET_FINGERTIP_OFFSET = 1031;
 
-    // Control commands.
+    // Control commands
     static constexpr uint16_t CMD_GRIP = 1;
     static constexpr uint16_t CMD_STOP = 8;
     static constexpr uint16_t CMD_GRIP_WITH_OFFSET = 16;
 
-    // Helper function to send a MODBUS request.
+    // Helper function to send a MODBUS request
     MB::ModbusResponse sendRequest(const MB::ModbusRequest &req);
 };
